Squash from #41, some notification codes are conflict;
Fixed typo

Small grammar and style fixes.

Actualization documentation

Update of russian version of documentation based on english version.

Add Promises section to the JS Guide.

Make it clearer that Promises is the way to go with JS.

Change incorrect string concatenation character

In the very first PHP code example, the string concat character is given as "+" in the code. It should be "." I think this is critical, because most users might copy-paste the first example and might get "0" if there was an exception, instead of the actual message.

Fix Links in Performance Guide

Also fixes bug when Swift is selected, since there are no Swift examples for these yet.

Facebook SDK Integration clarification

The Parse SDK does not automatically fetch Facebook data based on the read p

this was a typo

As evidenced by the same sentence later in the parse website (where I read this document) the line should read "just by" and not "just be".

Clarify subclassing instructions and difference between +load and +initialize.

Update objects.mdown

Fix typo in REST reference section

Fix spacing

get rid of ��

Remove ��

Minor grammar corrections in the Facebook user section.