Update queries.markdown "Queries on Array Values"
When querying on arrays, `whereKey=` needs to be used instead of `where=`. I validated this behaviour via the API Console, and CURL. It is also documented as an answer here: https://www.parse.com/questions/query-where-relation-contains-any-object-from-array.

FWIW, I think this behavior is unintuitive and would rather be able to use the `where` query key as with all other queries.