Windows installation (#463)
* Add dev-win npm script

This script should be used in place of dev npm script when user is running npm from a Windows command prompt.

Use dev-win when running command on a Windows command prompt. In Bash, the & operator forks and runs processes separately while in Windows MS-DOS, & just runs the command sequentally. Use start command to run commands in parallel

* Remove comment from package.json from last commit

* Add documentation for running docs on Windows

Add documentation for the dev-win npm script and how to correctly run the docs on Windows.