Document Async Current User (JS)
Adds a small note in the documentation regarding usage of `currentAsync()` when storage is asynchronous. This is noted lack of documentation in issue #500. 

What still stands is whether we should suggest usage of `currentAsync` or `currentUserAsync`. Since it seems the former was used to resolve the issue mentioned above I have chosen that for an additition the docs.